Restore Async lib, implement Mutex locking
