--- Software/src/lib/mathieucarbou-AsyncTCP/library.properties
@@ -1,10 +1,0 @@
-name=Async TCP
-includes=AsyncTCP.h
-version=3.3.1
-author=Me-No-Dev
-maintainer=Mathieu Carbou <mathieu.carbou@gmail.com>
-sentence=Async TCP Library for ESP32
-paragraph=Async TCP Library for ESP32
-category=Other
-url=https://github.com/mathieucarbou/AsyncTCP.git
-architectures=*
