+++ Software/src/lib/me-no-dev-AsyncTCP/library.properties
@@ -1,0 +1,9 @@
+name=AsyncTCP
+version=1.1.1
+author=Me-No-Dev
+maintainer=Me-No-Dev
+sentence=Async TCP Library for ESP32
+paragraph=Async TCP Library for ESP32
+category=Other
+url=https://github.com/me-no-dev/AsyncTCP
+architectures=*
